+++ src/main/java/com/acm/jwt/auth/service/AuthorizationService.java
@@ -1,0 +1,15 @@
+package com.acm.jwt.auth.service;
+
+import com.acm.app.user.client.domain.User;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthorizationService
+{
+
+    public User verifyUser(String username, String password)
+    {
+        return null;
+    }
+
+}