+++ src/main/java/com/acm/jwt/model/JwtRequest.java
@@ -1,0 +1,13 @@
+package com.acm.jwt.model;
+
+public class JwtRequest
+{
+    public String getUsername()
+    {
+        return null;
+    }
+    public String getPassword()
+    {
+        return null;
+    }
+}