+++ src/main/java/com/acm/jwt/model/JwtResponse.java
@@ -1,0 +1,12 @@
+package com.acm.jwt.model;
+
+public class JwtResponse
+{
+    private String token;
+
+
+    public JwtResponse(String token)
+    {
+        this.token = token;
+    }
+}