+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 # ChangeLog Template
 
-## Release - _--/--/---_ vX.X.X.X
+
+## Release - _--/--/---_ vX.X.X
+
+
    ### <ins>Added:</ins>
    ```diff
    + [name] - Description of information that was added goes here
@@ -20,7 +23,9 @@
    - [name] - Description of information that was added goes here
    ```
 ---
-## Release - 07/26/2020 v1.0.0.18
+
+## Release - 07/26/2020 v1.0.16
+
    ### <ins>Added:</ins>
    ```diff
    + [sambutler1017] - DqlBuilder class to read querys from acmsql file.
@@ -34,12 +39,16 @@
    ! [sambutler1017] - CodeOwners to include @llengel and @kwinborne
    ! [kwinborne] - Merged UserProfile and UserCredentials into Use
    ```
-## Release - 07/27/2020 v1.0.0.19
+
+## Release - 07/27/2020 v1.0.16
+
    ### <ins>Changed:</ins>
    ```diff
    ! [kwinborne] - Updated Readme.md to include the skeleton for a guide for beginners
    ```
-## Release - 07/28/2020 v1.0.0.20
+
+## Release - 07/28/2020 v1.0.16
+
    ### <ins>Added:</ins>
    ```diff
    + [kwinborne] - Test for UserClient
@@ -49,7 +58,9 @@
    ```diff
    ! [kwinborne] - Filled in UserService, UserClient, UserController with appropriate methods
    ```
-## Release - 07/28/2020 v1.0.0.21
+
+## Release - 07/28/2020 v1.0.16
+
    ### <ins>Added:</ins>
    ```diff
    + [kwinborne] - Created UserGetRequest class
@@ -59,3 +70,12 @@
    ! [kwinborne] - Modified user services to accomdate addition
    ! [kwinborne] - Cleaned up javadoc
    ```
+
+## Release - 07/28/2020 v1.0.16
+   ### <ins>Added:</ins>
+   ```diff
+   + [kwinborne] - Created Authentication Service class
+   + [kwinborne] - Added dummy classes for JWT model
+   + [kwinborne] - Created Authentication Controller class
+   ```
+